Add 'push_type' distro parameter to specify the build type to be used when a change is pushed to the repository.
The default is the old behavior of running the default
build type (typically 'ci'), but if you don't want a CI
build to happen for a particular distro configuration,
you can set this to None, and no CI scheduler will get
created.

Useful for when you have a separate SDK configuration for
a distro that you only want to run weekly, or through
a forced build.