ec2: reduce build_wait_timeout by default
With spot instances, it doesn't make sense to
try any more than one build at a time, since
the instance is apt to get shut down and reclaimed,
so default the wait timeout to zero in that case.

For normal on-demand instances, reduce the wait
timeout to five minutes, so there's still some
window for picking up a subsequent build.