Serialize ImageSet builds by default
Rather than trying to figure out how best to optimize
parallel ImageSet builds for variants of a distro,
default to serializing their execution. Builds may
actually end up being faster this way, if there is
sufficient sstate reuse between the variants.

A parallel_builds parameter is added to the
Distro() constructor to allow for having the
variants build in parallel on different workers.